+++ src/blocks/icon/edit.js
@@ -80,7 +80,7 @@ export default function IconEdit(props) {
 						className={`vk_icon_custombox_number`}
 						value={iconSize}
 						onChange={(value) =>
-							setAttributes({ iconSize: parseInt(value) })
+							setAttributes({ iconSize: parseFloat(value) })
 						}
 						type={'number'}
 					/>
@@ -127,7 +127,7 @@ export default function IconEdit(props) {
 						className={`vk_icon_custombox_number`}
 						value={iconMargin}
 						onChange={(value) =>
-							setAttributes({ iconMargin: parseInt(value) })
+							setAttributes({ iconMargin: parseFloat(value) })
 						}
 						type={'number'}
 					/>
